Populate initial kube-prometheus-stack values
Disable kubeControllerManager, kubeProxy and kubeScheduler because with
a mostly empty configuration the following alerts triggers on k3s:

 - KubeControllerManagerDown
 - KubeProxyDown
 - KubeSchedulerDown

Disable grafana: it will be eventually added as a separate HelmRelease
in order to being able to manage and update it more easily.

Start to initially retain metrics only for 24 hours and reserve just a
2Gi volume for them. Based on actual data it will be eventually
adjusted (it would be nice to have an higher retention but for the
moment better to start with small numbers).