Add slack-api-url secret
The secret was edited as-is (in clear-text) and then encrypted in-place via:

 sops --encrypt --in-place slack-api-url.yaml

(Given that we have a .sops.yaml file sops already pick up the right public key
to encrypt it without asking anything.)

This also start populating `secrets' directory for real. In order to
avoid adjusting kubeconform checks we can avoid to add a
`kustomization.yaml' because one will be enabled automatically by Flux
and `.sops.yaml' is already ignored by default so everything should
works.