Stop ignoring missing schemas and just ignore CRD
Ignoring missing schemas can ends up ignoring also a lot of resource
validation as it happened accidentally previously for Flux until commit
7bfa7ff.

Right now the only problematic resources are the
CustomResourceDefinition. Datree has a CRDs-catalog that can be used
in such cases that we should probably start using. For the moment
though, just skip them as it was until now.