Add victorialogs
Add VictoriaLogs configured as a single node.

Intended to replace Loki that is much more complex to manage and for
RPI use case probably less suitable.

Mostly the default configuration - VictoriaLogs do not need any special
configuration and defaults are sane - except the adjustment to use a 5Gi
volume (instead of default of 10Gi) and lower down the retention period
and adjust the retention disk space usage too according the volume size.

A corresponding serviceMonitor so that it is monitored by Prometheus is
also enabled.

In the first iteration promtail will be used to ingest logs, this will
be done in a separate commit later.
Probably we should investigate Vector - maybe deployed as a separate
Helm release - that it is used by default and can be enabled also in the
victoria-logs-single Helm chart.