Switch to use datreeio CRDs catalog
PodMonitor CRD is not present... switch to use datreeio CRDs catalog
that contains a lot of CRDs.

Please note that now kubeconform checks relies on the availability of
datreeio CRDs catalog... This is suboptimal.

We should find actual CRDs that we need and use and fetch/cache all of
them instead.