Add a telegram receiver for Alertmanager
Define another receiver via Telegram so that we have at least 3
different receivers.

Please note that it does not set any "parse_mode" because both
MarkdownV2 and HTML are very prone to error/injection.
I have tested with MarkdownV2 and for example "-" needs escaping and I
believe it is too difficult to have control of that based on the alert
labels and annotations.