Relax NodeMemoryHighUtilization
Redefine NodeMemoryHighUtilization alert in order to trigger when >95%
of the memory is used instead of the default of >90%.

That should be acceptable and it also seems that with that there are
still no Pod-s that ends up in OOM for the current workload.

Should be enough to silence #185 by still having the alert.